+++ app/Models/SystemSetting.php
@@ -2,15 +2,4 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class SystemSetting extends Model
-{
-    use HasFactory;
-
-    protected $fillable = [
-        'key',
-        'value',
-    ];
-}
+class SystemSetting extends \WebBlocks\Cms\Models\SystemSetting {}